+++ frontend/src/data/I18n/CreatePlotStepEight.ts
@@ -1,8 +1,8 @@
 export default {
   pt: {
     CreatePlotStepEight: {
-      title: 'Amostra 3',
-      subtitle: 'Repita o passso anterior, essa sera a ultima amostra',
+      title: 'Amostra 2',
+      subtitle: 'Repita o passso anterior, em outro ponto do talhão',
       sampleA: 'Quantidade total de grãos na planta A',
       sampleB: 'Quantidade total de grãos na planta B',
       samplePlaceholder: 'Digite a quantidade de grãos',
@@ -19,8 +19,8 @@ export default {
   },
   en: {
     CreatePlotStepEight: {
-      title: 'Sample 3',
-      subtitle: 'Repeat the previous step, this will be the last sample',
+      title: 'Sample 2',
+      subtitle: 'Repeat the previous step, at another point in the field',
       sampleA: 'Total amount of grains in plant A',
       sampleB: 'Total amount of grains in plant B',
       samplePlaceholder: 'Enter the amount of beans',
