+++ frontend/src/data/I18n/CreatePlotStepNine.ts
@@ -1,20 +1,38 @@
 export default {
   pt: {
     CreatePlotStepNine: {
-      title: 'Imagem das amostras',
-      subtitle: 'Tire uma foto de todas as plantas usadas nas amostras',
-      imagePlaceholder: 'Adicionar imagem',
-      imageUpdatePictureLabel: 'Alterar imagem',
-      finishButton: 'Finalizar'
+      title: 'Amostra 3',
+      subtitle: 'Repita o passso anterior, essa sera a ultima amostra',
+      sampleA: 'Quantidade total de grãos na planta A',
+      sampleB: 'Quantidade total de grãos na planta B',
+      samplePlaceholder: 'Digite a quantidade de grãos',
+      sampleDescription: 'Descrição (opcional)',
+      sampleDescriptionPlaceholder: 'Descreva aqui...',
+      buttonTitle: 'Continuar',
+      errors: {
+        grainsPlant: {
+          required: 'Quantidade de grãos é obrigatório',
+          min: 'Quantidade de grãos não pode ser "ZERO"'
+        }
+      }
     }
   },
   en: {
     CreatePlotStepNine: {
-      title: 'Picture of the photos',
-      subtitle: 'Take a photo of all the plants used in the samples',
-      imagePlaceholder: 'Add image',
-      imageUpdatePictureLabel: 'Change image',
-      finishButton: 'Finish'
+      title: 'Sample 3',
+      subtitle: 'Repeat the previous step, this will be the last sample',
+      sampleA: 'Total amount of grains in plant A',
+      sampleB: 'Total amount of grains in plant B',
+      samplePlaceholder: 'Enter the amount of beans',
+      sampleDescription: 'Description (optional)',
+      sampleDescriptionPlaceholder: 'Describe here...',
+      buttonTitle: 'Continue',
+      errors: {
+        grainsPlant: {
+          required: 'Grain quantity is required',
+          min: 'Grain quantity cannot be "ZERO"'
+        }
+      }
     }
   }
 };
