+++ frontend/src/routes/app.routes.tsx
@@ -9,13 +9,13 @@ import { useTheme } from 'styled-components';
 import { NavigatorProps } from '.';
 import Logo from '../assets/images/logo-dark.png';
 import { AppRoutesParams } from '../data/routes/app';
-import { CreatePlotStepEight } from '../screens/CreatePlot/CreatePlotStepEight';
+import { CreatePlotStepNine } from '../screens/CreatePlot/CreatePlotStepNine';
 import { CreatePlotStepFive } from '../screens/CreatePlot/CreatePlotStepFive';
 import { CreatePlotStepFour } from '../screens/CreatePlot/CreatePlotStepFour';
-import { CreatePlotStepNine } from '../screens/CreatePlot/CreatePlotStepNine';
-import { CreatePlotStepOne } from '../screens/CreatePlot/CreatePlotStepOne';
-import { CreatePlotStepSeven } from '../screens/CreatePlot/CreatePlotStepSeven';
 import { CreatePlotStepSix } from '../screens/CreatePlot/CreatePlotStepSix';
+import { CreatePlotStepOne } from '../screens/CreatePlot/CreatePlotStepOne';
+import { CreatePlotStepEight } from '../screens/CreatePlot/CreatePlotStepEight';
+import { CreatePlotStepSeven } from '../screens/CreatePlot/CreatePlotStepSeven';
 import { CreatePlotStepThree } from '../screens/CreatePlot/CreatePlotStepThree';
 import { CreatePlotStepTwo } from '../screens/CreatePlot/CreatePlotStepTwo';
 import { Home } from '../screens/Home';
