+++ frontend/src/screens/CreatePlot/CreatePlotStepNine/styles.ts
@@ -1,9 +1,5 @@
 import styled from 'styled-components/native';
-import {
-  RFFontSize,
-  RFHeight,
-  RFWidth
-} from '../../../utils/getResponsiveSizes';
+import { RFHeight, RFWidth } from '../../../utils/getResponsiveSizes';
 
 export const Container = styled.View`
   flex: 1;
@@ -13,10 +9,19 @@ export const Container = styled.View`
   background: ${({ theme }) => theme.colors.background_over};
 `;
 
-export const PictureContainer = styled.View`
+export const StepEightHelperImage = styled.Image`
+  width: 140px;
+  height: 140px;
+`;
+
+export const HelperImageContainer = styled.View`
+  width: 200px;
+  height: 200px;
   align-items: center;
   justify-content: center;
-  margin: ${RFHeight(24)}px auto;
+  background: ${({ theme }) => theme.colors.background_over};
+  border-radius: 100px;
+  margin: ${RFHeight(16)}px auto;
 `;
 
 export const FormContainer = styled.View`
@@ -31,12 +36,4 @@ export const FormContainer = styled.View`
 
 export const NextStepButton = styled.View`
   margin: ${RFHeight(24)}px 0;
-`;
-
-export const NoNetworkMessage = styled.Text`
-  color: ${({ theme }) => theme.colors.text};
-  font-size: ${RFFontSize(16)}px;
-  font-family: ${({ theme }) => theme.fonts.regular};
-  text-align: center;
-  margin: ${RFHeight(48)}px 0;
 `;
