+++ frontend/src/screens/CreatePlot/CreatePlotStepSeven/index.tsx
@@ -3,12 +3,12 @@ import React, { useEffect } from 'react';
 import { FieldValues, useForm } from 'react-hook-form';
 import { ScrollView } from 'react-native';
 import * as yup from 'yup';
-import StepSeven from '../../../assets/plot-steps-images/StepSample.png';
+import StepSix from '../../../assets/plot-steps-images/StepSample.png';
 import { Button } from '../../../components/Button';
 import { StepIndicator } from '../../../components/StepIndicator';
 import { TextInput } from '../../../components/TextInput';
 import Title from '../../../components/Title';
-import { CreatePlotStepSevenScreenRouteProps } from '../../../data/routes/app';
+import { CreatePlotStepSixScreenRouteProps } from '../../../data/routes/app';
 import { useSample } from '../../../hooks/useSample';
 import { translate } from '../../../data/I18n';
 import {
@@ -16,7 +16,7 @@ import {
   FormContainer,
   HelperImageContainer,
   NextStepButton,
-  StepSevenHelperImage
+  StepSixHelperImage
 } from './styles';
 
 const userLogin = yup.object().shape({
@@ -30,9 +30,9 @@ const userLogin = yup.object().shape({
     .min(1, 'Quantidade de grãos não pode ser "ZERO"')
 });
 
-export const CreatePlotStepSeven: React.FC<
-  CreatePlotStepSevenScreenRouteProps
-> = ({ navigation }) => {
+export const CreatePlotStepSeven: React.FC<CreatePlotStepSixScreenRouteProps> = ({
+  navigation
+}) => {
   const { saveStep, getPersistedData } = useSample();
 
   const {
@@ -47,25 +47,25 @@ export const CreatePlotStepSeven: React.FC<
   useEffect(() => {
     getPersistedData().then(data => {
       if (data) {
-        setValue('grainsPlant1', data?.plantB?.grainsPlant1?.toString() || '');
-        setValue('grainsPlant2', data?.plantB?.grainsPlant2?.toString() || '');
-        setValue('description', data?.plantB?.description || '');
+        setValue('grainsPlant1', data?.plantA?.grainsPlant1?.toString() || '');
+        setValue('grainsPlant2', data?.plantA?.grainsPlant2?.toString() || '');
+        setValue('description', data?.plantA?.description || '');
       }
     });
   }, [getPersistedData, setValue]);
 
-  const handleSubmitStepSeven = (data: FieldValues) => {
+  const handleSubmitStepSix = (data: FieldValues) => {
     const sample: any = {
-      plantB: {
+      plantA: {
         grainsPlant1: data.grainsPlant1,
         grainsPlant2: data.grainsPlant2
       }
     };
     if (data?.description) {
-      sample.plantB.description = data.description;
+      sample.plantA.description = data.description;
     }
     saveStep(sample);
-    navigation.navigate('CreatePlotStepEight');
+    navigation.navigate('CreatePlotStepSeven');
   };
 
   return (
@@ -75,10 +75,10 @@ export const CreatePlotStepSeven: React.FC<
           title={translate('CreatePlotStepSeven.title')}
           subtitle={translate('CreatePlotStepSeven.subtitle')}
         />
-        <StepIndicator step={1} indicator={5} />
+        <StepIndicator step={1} indicator={4} />
         <FormContainer>
           <HelperImageContainer>
-            <StepSevenHelperImage source={StepSeven} resizeMode="contain" />
+            <StepSixHelperImage source={StepSix} resizeMode="contain" />
           </HelperImageContainer>
           <TextInput
             label="CreatePlotStepSeven.sampleA"
@@ -107,8 +107,8 @@ export const CreatePlotStepSeven: React.FC<
           />
           <NextStepButton>
             <Button
-              title={translate('CreatePlotStepSeven.buttonTitle')}
-              onPress={handleSubmit(handleSubmitStepSeven)}
+              title="Continuar"
+              onPress={handleSubmit(handleSubmitStepSix)}
             />
           </NextStepButton>
         </FormContainer>
