+++ frontend/src/screens/CreatePlot/CreatePlotStepSix/styles.ts
@@ -1,5 +1,9 @@
 import styled from 'styled-components/native';
-import { RFHeight, RFWidth } from '../../../utils/getResponsiveSizes';
+import {
+  RFFontSize,
+  RFHeight,
+  RFWidth
+} from '../../../utils/getResponsiveSizes';
 
 export const Container = styled.View`
   flex: 1;
@@ -9,19 +13,10 @@ export const Container = styled.View`
   background: ${({ theme }) => theme.colors.background_over};
 `;
 
-export const StepSixHelperImage = styled.Image`
-  width: 140px;
-  height: 140px;
-`;
-
-export const HelperImageContainer = styled.View`
-  width: 200px;
-  height: 200px;
+export const PictureContainer = styled.View`
   align-items: center;
   justify-content: center;
-  background: ${({ theme }) => theme.colors.background_over};
-  border-radius: 100px;
-  margin: ${RFHeight(16)}px auto;
+  margin: ${RFHeight(24)}px auto;
 `;
 
 export const FormContainer = styled.View`
@@ -36,4 +31,12 @@ export const FormContainer = styled.View`
 
 export const NextStepButton = styled.View`
   margin: ${RFHeight(24)}px 0;
+`;
+
+export const NoNetworkMessage = styled.Text`
+  color: ${({ theme }) => theme.colors.text};
+  font-size: ${RFFontSize(16)}px;
+  font-family: ${({ theme }) => theme.fonts.regular};
+  text-align: center;
+  margin: ${RFHeight(48)}px 0;
 `;
